매니패스트 주석 버그 수정 및 메인과 맵 구분
